feat(searchBar): just the UI
