+++ example.env
@@ -5,3 +5,4 @@ OAUTH_CLIENT_ID=2c29d26a0e4bd9ee8c35
 OAUTH_CLIENT_SECRET=SECRET
 OAUTH_CALLBACK_URL=http://localhost:3001/auth/callback
 SERVER_PORT=3001
+PG_CONNECTION_STRING="postgres://postgres:Aa123456@localhost:5432/jobs"
