+++ server/db.js
@@ -2,13 +2,9 @@
 require('dotenv').config();
 const { Pool } = require('pg')
 
-const user = process.env.PG_USERNAME || "postgres";
-const password = process.env.PG_PASSWORD || "postgres";
-const host = process.env.PG_HOST || "localhost";
-const database = process.env.PG_DB || "jobs";
-const port = process.env.PG_PORT || 5432;
+const connectionString = process.env.PG_CONNECTION_STRING || `postgres://postgres:Aa123456@localhost:5432/jobs`;
 
-const pool = new Pool({ user, host, database, password, port });
+const pool = new Pool({connectionString});
 
 module.exports = {
   query: (text, params) => pool.query(text, params),
